src/main/java/spring/noticeboard/service/login/LoginService.java
package spring.noticeboard.service.login;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import spring.noticeboard.domain.member.Member;
import spring.noticeboard.repository.member.MemberRepository;

import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpSession;

@Service
@RequiredArgsConstructor
public class LoginService {

    private final MemberRepository memberRepository;

    public Member login(String loginId, String password) {
        return memberRepository.findByLoginId(loginId)
                .filter(m -> m.getPassword().equals(password))
                .orElse(null);
    }

    public void logout(HttpServletRequest request) {
        HttpSession session = request.getSession(false);

        if (session != null) {
            session.invalidate();
        }
    }
}
